dashboard/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-storage==2.18.2
google-cloud-run==0.10.7
google-cloud-scheduler==2.15.0
google-cloud-secret-manager==2.21.1
pyyaml==6.0.2
python-multipart==0.0.12
pydantic==2.9.2
